+++ test/common/delegate.cpp
@@ -21,8 +21,8 @@ class Helper {
 
 TEST_CASE("Delegate works as expected") {
     Helper helper;
-    helper.on([&helper]() {
-        helper.on([&helper]() {});
+    helper.on([&]() {
+        helper.on([&]() {});
     });
     helper.emit();
 }